backend/src/main/java/greenpulse/ecocrops/ecocrops/python/requirements.txt
requests
pandas
numpy
pickle-mixin
python-dotenv
google-generativeai
absl-py
